pyproject.toml
[project]
name = "herecrm"
requires-python = ">=3.12"
version = "0.1.0"
description = "A CRM right in Whatsapp!"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "requests",
    "httpx",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "sqlalchemy",
    "aiosqlite",
    "openai",
    "pytest",
    "pytest-asyncio",
    "PyYAML",
    "streamlit>=1.53.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "alembic>=1.18.1",
    "python-multipart",
    "boto3",
    "twilio>=9.9.1",
    "weasyprint",
    "jinja2",
    "spec-kitty-cli>=0.11.1",
    "stripe>=14.2.0",
    "python-quickbooks>=0.9.1",
    "apscheduler",
]

# Optional dependencies for QuickBooks integration:
# pysqlcipher3: Requires system sqlcipher library
# Ubuntu/Debian: sudo apt-get install sqlcipher
# Then: pip install pysqlcipher3

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "--ignore=tests/reproduce_issue.py --ignore=tests/reproduce_spatial_filtering_behavior.py --ignore=scripts/"
testpaths = ["tests"]

[tool.pyrefly]
import-roots = ["."]
search-paths = ["."]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
